+++ rpc/backend/node_info.go
@@ -33,10 +33,11 @@ import (
 func (b *Backend) Accounts() ([]common.Address, error) {
 	addresses := make([]common.Address, 0) // return [] instead of nil if empty
 
-	if b.cfg.JSONRPC.Enable && !b.cfg.JSONRPC.AllowInsecureUnlock {
-		b.logger.Error("Account unlock with HTTP access is forbidden")
-		return addresses, fmt.Errorf("Account unlock with HTTP access is forbidden!")
+	if !b.cfg.JSONRPC.AllowInsecureUnlock {
+		b.logger.Debug("account unlock with HTTP access is forbidden")
+		return addresses, fmt.Errorf("account unlock with HTTP access is forbidden!")
 	}
+
 	infos, err := b.clientCtx.Keyring.List()
 	if err != nil {
 		return addresses, err
@@ -82,8 +83,8 @@ func (b *Backend) Syncing() (interface{}, error) {
 
 // SetEtherbase sets the etherbase of the miner
 func (b *Backend) SetEtherbase(etherbase common.Address) bool {
-	if b.cfg.JSONRPC.Enable && !b.cfg.JSONRPC.AllowInsecureUnlock {
-		b.logger.Error("Account unlock with HTTP access is forbidden")
+	if !b.cfg.JSONRPC.AllowInsecureUnlock {
+		b.logger.Error("account unlock with HTTP access is forbidden")
 		return false
 	}
 
@@ -228,10 +229,11 @@ func (b *Backend) ImportRawKey(privkey, password string) (common.Address, error)
 func (b *Backend) ListAccounts() ([]common.Address, error) {
 	addrs := []common.Address{}
 
-	if b.cfg.JSONRPC.Enable && !b.cfg.JSONRPC.AllowInsecureUnlock {
-		b.logger.Error("Account unlock with HTTP access is forbidden")
-		return addrs, fmt.Errorf("Account unlock with HTTP access is forbidden!")
+	if !b.cfg.JSONRPC.AllowInsecureUnlock {
+		b.logger.Error("account unlock with HTTP access is forbidden")
+		return addrs, fmt.Errorf("account unlock with HTTP access is forbidden!")
 	}
+
 	list, err := b.clientCtx.Keyring.List()
 	if err != nil {
 		return nil, err